Use the head branch of a Pull Request as a fallback for config file
A lot of times, maintainers try to open a Pull Request containing the .pep8speaks.yml to check if the configurations
are working or not. In those situations, there will not be any configurations on the master branch. Hence it makes sense
to use the head branch of the Pull Request as a source for the config file.

Closes #55