Fix "single comment" behavior for non-official deployments
If updates are made to a pull request and a comment from the bot
already exists, that comment should be updated; a new comment should
not be added. However, this functionality is broken for non-official
deployments -- ie, deployments using a GitHub user other than
"pep8speaks". This commit fixes this by using the BOT_PASSWORD env var
to check for an existing comment, rather than hardcoding the user ID.